+++ src/library/renderhud/RenderHUD_GL.cpp
@@ -87,7 +87,7 @@ RenderHUD_GL::~RenderHUD_GL() {
     fini();
 }
 
-void RenderHUD_GL::init()
+void RenderHUD_GL::init(bool stateGLES)
 {
     if (texture == 0) {
         LINK_NAMESPACE(glGenTextures, "GL");
@@ -189,7 +189,26 @@ void RenderHUD_GL::init()
                 TexCoord = aTexCoord;
             }
         )";
-    	orig::glShaderSource(vertexShaderID, 1, &vertexShaderSource , NULL);
+
+        const char *vertexShaderSourceES = R"(
+            #version 100
+            attribute vec3 aPos;
+            attribute vec2 aTexCoord;
+            varying vec2 TexCoord;
+        
+            void main()
+            {
+                gl_Position = vec4(aPos, 1.0);
+                TexCoord = aTexCoord;
+            }
+        )";
+
+        if (stateGLES) {
+            orig::glShaderSource(vertexShaderID, 1, &vertexShaderSourceES , NULL);
+        }
+        else {
+            orig::glShaderSource(vertexShaderID, 1, &vertexShaderSource , NULL);            
+        }
     	orig::glCompileShader(vertexShaderID);
         
         
@@ -212,7 +231,26 @@ void RenderHUD_GL::init()
                 gl_FragColor = texture2D(ourTexture, TexCoord);
             }
         )";
-    	orig::glShaderSource(fragmentShaderID, 1, &fragmentShaderSource , NULL);
+
+        const char *fragmentShaderSourceES = R"(
+            #version 100
+            precision mediump float;
+            uniform sampler2D ourTexture;
+            varying vec2 TexCoord;
+        
+            void main()
+            {
+                gl_FragColor = texture2D(ourTexture, TexCoord);
+            }
+        )";
+        
+        if (stateGLES) {
+            orig::glShaderSource(fragmentShaderID, 1, &fragmentShaderSourceES , NULL);
+        }
+        else {
+            orig::glShaderSource(fragmentShaderID, 1, &fragmentShaderSource , NULL);
+        }
+
     	orig::glCompileShader(fragmentShaderID);
         
         /* Check the fragment shader */
@@ -273,7 +311,7 @@ void RenderHUD_GL::fini()
 
 void RenderHUD_GL::renderSurface(std::unique_ptr<SurfaceARGB> surf, int x, int y)
 {
-    RenderHUD_GL::init();
+    RenderHUD_GL::init(isGLES);
 
     LINK_NAMESPACE(glEnable, "GL");
     LINK_NAMESPACE(glDisable, "GL");